Close #88 - Simplify import
