+++ cats-effect/src/main/scala/loggerf/cats/LoggerA.scala
@@ -2,10 +2,8 @@ package loggerf.cats
 
 import cats._
 import cats.implicits._
-
 import effectie.cats.EffectConstructor
-
-import loggerf.Logger
+import loggerf.logger.Logger
 
 trait LoggerA[F[_]] {
