+++ cats-effect/src/test/scala/loggerf/cats/LoggerEitherSpec.scala
@@ -3,14 +3,11 @@ package loggerf.cats
 import cats._
 import cats.implicits._
 import cats.effect._
-
 import effectie.Effectful._
 import effectie.cats.EffectConstructor
-
 import hedgehog._
 import hedgehog.runner._
-
-import loggerf.LoggerForTesting
+import loggerf.logger.LoggerForTesting
 
 /**
  * @author Kevin Lee
