+++ cats-effect/src/test/scala/loggerf/cats/LoggerOptionSpec.scala
@@ -2,14 +2,11 @@ package loggerf.cats
 
 import cats._
 import cats.effect._
-
 import effectie.Effectful._
 import effectie.cats.EffectConstructor
-
 import hedgehog._
 import hedgehog.runner._
-
-import loggerf.LoggerForTesting
+import loggerf.logger.LoggerForTesting
 
 /**
  * @author Kevin Lee
