+++ core/src/main/scala/loggerf/logger/Logger.scala
@@ -1,4 +1,4 @@
-package loggerf
+package loggerf.logger
 
 /**
  * @author Kevin Lee
