+++ core/src/main/scala/loggerf/syntax/package.scala
@@ -1,0 +1,9 @@
+package loggerf
+
+/**
+ * @author Kevin Lee
+ * @since 2020-08-01
+ */
+package object syntax
+  extends LogSyntax
+     with LeveledMessageSyntax
