+++ log4j/src/main/scala/loggerf/logger/Log4jLogger.scala
@@ -1,4 +1,4 @@
-package loggerf
+package loggerf.logger
 
 import scala.reflect.ClassTag
