+++ scalaz-effect/src/main/scala/loggerf/scalaz/LoggerOptionT.scala
@@ -1,10 +1,8 @@
 package loggerf.scalaz
 
 import scalaz._
-
 import effectie.scalaz.EffectConstructor
-
-import loggerf.Logger
+import loggerf.logger.Logger
 
 trait LoggerOptionT[F[_]] {
