+++ scalaz-effect/src/main/scala/loggerf/scalaz/package.scala
@@ -1,0 +1,7 @@
+package loggerf
+
+/**
+ * @author Kevin Lee
+ * @since 2020-08-01
+ */
+package object scalaz extends Logful
