+++ scalaz-effect/src/test/scala/loggerf/scalaz/LogSpec.scala
@@ -3,15 +3,12 @@ package loggerf.scalaz
 import scalaz._
 import Scalaz._
 import scalaz.effect._
-
 import effectie.Effectful._
 import effectie.scalaz.EffectConstructor
-
 import hedgehog._
 import hedgehog.runner._
-
-import loggerf.LoggerForTesting
-import loggerf.scalaz.Log.LeveledMessage._
+import loggerf.logger.LoggerForTesting
+import loggerf.syntax._
 
 /**
  * @author Kevin Lee
