+++ scalaz-effect/src/test/scala/loggerf/scalaz/LoggerEitherSpec.scala
@@ -3,14 +3,11 @@ package loggerf.scalaz
 import scalaz._
 import Scalaz._
 import scalaz.effect._
-
 import effectie.Effectful._
 import effectie.scalaz.EffectConstructor
-
 import hedgehog._
 import hedgehog.runner._
-
-import loggerf.LoggerForTesting
+import loggerf.logger.LoggerForTesting
 
 /**
  * @author Kevin Lee
