+++ scalaz-effect/src/test/scala/loggerf/scalaz/LoggerOptionSpec.scala
@@ -2,14 +2,11 @@ package loggerf.scalaz
 
 import scalaz._
 import scalaz.effect._
-
 import effectie.Effectful._
 import effectie.scalaz.EffectConstructor
-
 import hedgehog._
 import hedgehog.runner._
-
-import loggerf.LoggerForTesting
+import loggerf.logger.LoggerForTesting
 
 /**
  * @author Kevin Lee
