+++ slf4j/src/main/scala/loggerf/logger/Slf4JLogger.scala
@@ -1,4 +1,4 @@
-package loggerf
+package loggerf.logger
 
 import scala.reflect.ClassTag
